pyproject.toml
[project]
name = "sasiki"
version = "0.1.0"
description = "观察用户操作，自动生成可复用的工作流 Agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"}
]
keywords = ["workflow", "automation", "agent", "multimodal", "screen-recording"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    # LLM
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    # Video/Image Processing
    "opencv-python>=4.9.0",
    "pillow>=10.2.0",
    "imagehash>=4.3.1",
    "scikit-image>=0.22.0",
    # macOS Screen Capture
    "pyobjc-core>=10.1; sys_platform == 'darwin'",
    "pyobjc-framework-AVFoundation>=10.1; sys_platform == 'darwin'",
    "pyobjc-framework-CoreMedia>=10.1; sys_platform == 'darwin'",
    "pyobjc-framework-Vision>=10.1; sys_platform == 'darwin'",
    # Automation
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "pynput>=1.7.6",
    # Utils
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic-yaml>=1.2.0",
    "aiofiles>=23.2.1",
    "playwright>=1.58.0",
    "mcp>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
sasiki = "sasiki.cli:main"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "sasiki.cli",
    "sasiki.analyzer.*",
    "sasiki.recorder.*",
]
ignore_errors = true

[tool.pytest.ini_options]
pythonpath = ["src"]
